Added home screen

+++ app/src/main/java/com/kotlinandroiddemo/api/model/Photo.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 /**
  * Created by Deven on 10/20/17.
  */
-data class Photo(@SerializedName("id") val id: Int, @SerializedName("owner") val owner: String,
+data class Photo(@SerializedName("id") val id: Long, @SerializedName("owner") val owner: String,
                  @SerializedName("secret") val secret: String, @SerializedName("server") val server: String,
                  @SerializedName("farm") val farm: Int, @SerializedName("title") val title: String,
                  @SerializedName("public") val public: Int, @SerializedName("friend") val friend: Int,
